Simplify Slack::EventSubscriptionsController
